June 4

+++ Relation_Info/views.py
@@ -6,13 +6,19 @@ from pyecharts.charts import Bar
 from pyecharts.charts import Map
 from pyecharts.faker import Faker
 
-from Relation_Info.models import Company_Branch_Delivery
+from Relation_Info.models import Company_Branch_Delivery, Supplier_Goods_Info
 from Entity_Info.models import User_Info, Branch_Info, Storage_Info, Goods_Info
 from datetime import date
 
+
 # Create your views here.
 class Record(object):
-    count = 0  # Record.count是静态变量
+    delivery_id = 0
+    bill = 0
+    delivery_request = 0
+    SKU = 0
+    Supplier_Goods_Info_Request = 0
+    Supplier_Goods_Info_Goods = 0
 
 
 def company_branch_delivery(request):
@@ -20,23 +26,32 @@ def company_branch_delivery(request):
     Edit_id = request.GET.get("Edit_id", '')
     Page_id = request.GET.get('page', 1)
     if Edit_id:
-        Record.count = Edit_id
+        Record.delivery_id = Edit_id
 
     # 删除
     Delete_id = request.GET.get("Delete_id", '')
-    if Delete_id:
+    if Delete_id and Company_Branch_Delivery.objects.filter(goods_delivery_id=Delete_id):
         Company_Branch_Delivery.objects.get(goods_delivery_id=Delete_id).delete()
 
     # 删完再查询
     Max_id = max([item['goods_delivery_id'] for item in Company_Branch_Delivery.objects.values('goods_delivery_id')])
     bill_lists = Company_Branch_Delivery.objects.order_by('-goods_delivery_id')
+    if Page_id != 1:
+        bill_lists = Record.bill
+        request.POST = Record.delivery_request
+    else:
+        Record.delivery_request = request.POST
+        Record.bill = bill_lists
+
+    bill_lists_len = len(bill_lists)
     paginator = Paginator(bill_lists, 3)
     bill_lists = paginator.page(Page_id)
+
     if request.GET.get('Reset'):
         return render(request, 'Company_Branch_Delivery.html',
                       {'bill_lists': bill_lists, 'Dist': District_list,
-                       'New_id': Max_id + 1, 'bill_len': len(bill_lists), 'paginator': paginator})
-    # if request.method == 'POST':
+                       'New_id': Max_id + 1, 'bill_len': bill_lists_len, 'paginator': paginator})
+
     # 修改
     Edit_goods_delivery_send_time = request.POST.get("edit_goods_delivery_send_time", '')
     Edit_goods_delivery_receive_time = request.POST.get("edit_goods_delivery_receive_time", '')
@@ -47,26 +62,35 @@ def company_branch_delivery(request):
     Edit_goods_delivery_staff_id = request.POST.get("edit_goods_delivery_staff_id", '')
 
     if Edit_goods_delivery_send_time:
-        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.count).update(
+        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.delivery_id).update(
             goods_delivery_send_time=Edit_goods_delivery_send_time)
     if Edit_goods_delivery_receive_time:
-        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.count).update(
+        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.delivery_id).update(
             goods_delivery_receive_time=Edit_goods_delivery_receive_time)
     if Edit_goods_num:
-        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.count).update(
+        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.delivery_id).update(
             goods_num=Edit_goods_num)
     if Edit_delivery_branch_id:
-        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.count).update(
+        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.delivery_id).update(
             delivery_branch_id_id=Edit_delivery_branch_id)
     if Edit_delivery_storage_id:
-        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.count).update(
+        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.delivery_id).update(
             delivery_storage_id_id=Edit_delivery_storage_id)
     if Edit_delivery_goods_id:
-        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.count).update(
+        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.delivery_id).update(
             delivery_goods_id_id=Edit_delivery_goods_id)
     if Edit_goods_delivery_staff_id:
-        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.count).update(
+        Company_Branch_Delivery.objects.filter(goods_delivery_id=Record.delivery_id).update(
             goods_delivery_staff_id_id=Edit_goods_delivery_staff_id)
+
+    Edit_id_branch_id = ''
+    Edit_id_branch_district = ''
+    Edit_object = ''
+    if Edit_id and Company_Branch_Delivery.objects.filter(goods_delivery_id=Edit_id):
+        Edit_id_branch_id = Branch_Info.objects.get(
+            branch_id=Company_Branch_Delivery.objects.get(goods_delivery_id=Edit_id).delivery_branch_id_id)
+        Edit_id_branch_district = Edit_id_branch_id.branch_district
+        Edit_object = Company_Branch_Delivery.objects.get(goods_delivery_id=Edit_id)
 
     # 搜索
     Search_Goods_delivery_id = request.POST.get("Search_Goods_delivery_id")
@@ -75,16 +99,17 @@ def company_branch_delivery(request):
     Search_Staff_id = request.POST.get("Search_Staff_id")
     Search_Branch_id = request.POST.get("Search_Branch_id", 'All')
     Msg = {'msg_1': '', 'msg_2': '', 'msg_3': '', 'msg_4': '', 'msg_5': ''}
-    if Search_Staff_id and not User_Info.objects.filter(user_account__contains=Search_Staff_id):
-        Msg['msg_1'] = '员工号无此字段'
     if Search_Goods_delivery_id and not Company_Branch_Delivery.objects.filter(
             goods_delivery_id__contains=Search_Goods_delivery_id):
-        Msg['msg_2'] = '单号无此字段'
+        Msg['msg_1'] = '无此单号'
+    if Search_Staff_id and not User_Info.objects.filter(user_account__contains=Search_Staff_id):
+        Msg['msg_2'] = '无此员工号'
+
     if Msg['msg_1'] or Msg['msg_2']:
         return render(request, 'Company_Branch_Delivery.html',
                       {'bill_lists': bill_lists, 'Dist': District_list, 'Msg': Msg,
                        'New_id': Max_id + 1, 'Edit_id': Edit_id, 'All_request': request.POST,
-                       'bill_len': len(bill_lists), 'paginator': paginator})
+                       'bill_len': bill_lists_len, 'paginator': paginator})
 
     condition_dict = {}
     if not Search_Time_begin:
@@ -108,20 +133,21 @@ def company_branch_delivery(request):
     New_delivery_storage_id = request.POST.get("new_delivery_storage_id", '')
     New_delivery_goods_id = request.POST.get("new_delivery_goods_id", '')
     New_goods_delivery_staff_id = request.POST.get("new_goods_delivery_staff_id", '')
-    if New_delivery_storage_id and not Storage_Info.objects.filter(storage_id=New_delivery_storage_id):
-        Msg['msg_3'] = '仓库不存在'
     if New_delivery_goods_id and not Goods_Info.objects.filter(goods_id=New_delivery_goods_id):
-        Msg['msg_4'] = '商品不存在'
+        Msg['msg_3'] = '商品编号不存在'
+    if New_delivery_storage_id and not Storage_Info.objects.filter(storage_id=New_delivery_storage_id):
+        Msg['msg_4'] = '仓库编号不存在'
     if New_goods_delivery_staff_id and not User_Info.objects.filter(user_account=New_goods_delivery_staff_id):
         Msg['msg_5'] = '员工不存在'
-    elif New_goods_delivery_staff_id and not User_Info.objects.get(user_account=New_goods_delivery_staff_id).user_dep == '物流部门':
-        Msg['msg_5'] = '员工不属物流部门'
+    elif New_goods_delivery_staff_id and not User_Info.objects.get(
+            user_account=New_goods_delivery_staff_id).user_dep == '物流部门':
+        Msg['msg_5'] = '此员工不属物流部门'
 
     if Msg['msg_3'] or Msg['msg_4'] or Msg['msg_5']:
         return render(request, 'Company_Branch_Delivery.html',
                       {'bill_lists': bill_lists, 'Dist': District_list, 'Msg': Msg,
                        'New_id': Max_id + 1, 'Edit_id': Edit_id, 'All_request': request.POST,
-                       'bill_len': len(bill_lists), 'paginator': paginator})
+                       'bill_len': bill_lists_len, 'paginator': paginator})
     if New_goods_delivery_send_time and New_delivery_branch_id and New_delivery_storage_id and New_delivery_goods_id and New_goods_delivery_staff_id:
         Company_Branch_Delivery.objects.create(goods_delivery_id=Max_id + 1,
                                                goods_delivery_send_time=New_goods_delivery_send_time,
@@ -172,22 +198,24 @@ def company_branch_delivery(request):
         elif Sort_condition(Staff_id_Sort):
             sort_list.append(Sort_condition(Staff_id_Sort))
     bill_lists = bill_lists.order_by(*sort_list)
-    bill_lists_len = len(bill_lists)
-    # print('POST内容: ', request.POST)
 
+    # print('POST内容: ', request.POST)
+    if Page_id != 1:
+        bill_lists = Record.bill
+        request.POST = Record.delivery_request
+    else:
+        Record.delivery_request = request.POST
+        Record.bill = bill_lists
+
+    bill_lists_len = len(bill_lists)
     paginator = Paginator(bill_lists, 3)
     bill_lists = paginator.page(Page_id)
-    Edit_id_branch_id = ''
-    Edit_id_branch_district = ''
-    if Edit_id:
-        Edit_id_branch_id = Branch_Info.objects.get(
-            branch_id=Company_Branch_Delivery.objects.get(goods_delivery_id=Edit_id).delivery_branch_id_id)
-        Edit_id_branch_district = Edit_id_branch_id.branch_district
+
     return render(request, 'Company_Branch_Delivery.html',
                   {'bill_lists': bill_lists, 'Dist': District_list, 'New_id': Max_id + 1,
                    'Edit_id': Edit_id, 'All_request': request.POST, 'bill_len': bill_lists_len,
                    'Edit_id_branch_id': Edit_id_branch_id, 'Edit_id_branch_district': Edit_id_branch_district,
-                   'paginator': paginator})
+                   'Edit_object': Edit_object, 'paginator': paginator})
 
 
 def Sort_condition(mode):
@@ -223,8 +251,17 @@ def Sort_condition(mode):
         return 'goods_delivery_staff_id'
     elif mode == 'staff_id_desc':
         return '-goods_delivery_staff_id'
+    elif mode == 'SKU_incr':
+        return 'SKU'
+    elif mode == 'SKU_desc':
+        return '-SKU'
+    elif mode == 'order_price_incr':
+        return 'order_price'
+    elif mode == 'order_price_desc':
+        return '-order_price'
     else:
         return 0
+
 
 # def company_branch_delivery_show(request):
 #     if request.POST.get('Show_state'):
@@ -250,3 +287,118 @@ def Sort_condition(mode):
 #                       {'bill_lists': bill_lists, 'Dist': District_list, 'New_id': Max_id + 1,
 #                        'Edit_id': Edit_id, 'data': c.render_embed(),
 #                        'All_request': request.POST, 'bill_len': bill_lists_len})
+
+def supplier_goods_info(request):
+    Edit_id = request.GET.get("Edit_id", '')
+    Page_id = request.GET.get('page', 1)
+    if Edit_id:
+        Record.count = Edit_id
+
+    # 删除
+    Delete_id = request.GET.get("Delete_id", '')
+    if Delete_id and Supplier_Goods_Info.objects.filter(SKU=Delete_id):
+        Supplier_Goods_Info.objects.get(SKU=Delete_id).delete()
+
+    # 删完再查询
+    goods_lists = Supplier_Goods_Info.objects.order_by('-SKU')
+    if Page_id != 1:
+        goods_lists = Record.Supplier_Goods_Info_Goods
+        request.POST = Record.Supplier_Goods_Info_Request
+    else:
+        Record.Supplier_Goods_Info_Request = request.POST
+        Record.Supplier_Goods_Info_Goods = goods_lists
+    goods_len = len(goods_lists)
+    paginator = Paginator(goods_lists, 10)
+    goods_lists = paginator.page(Page_id)
+    if request.GET.get('Reset'):
+        return render(request, 'Supplier_Goods_Info.html',
+                      {'goods_lists': goods_lists, 'goods_len': goods_len, 'paginator': paginator})
+    # if request.method == 'POST':
+
+    # 修改
+    Edit_SKU = request.POST.get("edit_SKU", '')
+    Edit_order_price = request.POST.get("edit_order_price", '')
+    if Edit_SKU:
+        Supplier_Goods_Info.objects.filter(SKU=Record.SKU).update(
+            SKU=Edit_SKU)
+    if Edit_order_price:
+        Supplier_Goods_Info.objects.filter(SKU=Record.SKU).update(
+            order_price=Edit_order_price)
+    Edit_object = ''
+    if Edit_id and Supplier_Goods_Info.objects.filter(SKU=Edit_id):
+        Edit_object = Supplier_Goods_Info.objects.get(SKU=Edit_id)
+    # 查询
+    Search_SKU = request.POST.get("Search_SKU")
+    Search_price = request.POST.get("Search_price")
+    condition_dict = {}
+    Search_price_low_bound = '0'
+    Search_price_up_bound = '99999'
+    Find_Index = str(Search_price).find('-')
+    if Search_price and Find_Index == -1:
+        Search_price_low_bound = Search_price
+        Search_price_up_bound = Search_price
+    elif Search_price and Find_Index == 0:
+        Search_price_up_bound = str(Search_price)[Find_Index + 1:]
+    elif Search_price and Find_Index == len(Search_price) - 1:
+        Search_price_low_bound = str(Search_price)[:Find_Index]
+    elif Search_price:
+        Search_price_low_bound = str(Search_price)[:Find_Index]
+        Search_price_up_bound = str(Search_price)[Find_Index + 1:]
+    if Search_SKU:
+        condition_dict['SKU__contains'] = Search_SKU
+    Msg = {'msg_1': '', 'msg_2': ''}
+    if Search_SKU and not Supplier_Goods_Info.objects.filter(SKU__contains=Search_SKU):
+        Msg['msg_1'] = '店内码无此数据项'
+    if Msg['msg_1']:
+        return render(request, 'Supplier_Goods_Info.html',
+                      {'goods_lists': goods_lists, 'Msg': Msg,
+                       'Edit_id': Edit_id, 'All_request': request.POST,
+                       'goods_len': goods_len, 'paginator': paginator})
+    condition_dict['order_price__range'] = [Search_price_low_bound, Search_price_up_bound]
+    goods_lists = Supplier_Goods_Info.objects.filter(**condition_dict)
+
+    # 新增
+    New_SKU = request.POST.get("new_SKU", '')
+    New_order_price = request.POST.get("new_order_price", '')
+    goods_len = len(paginator.page(Page_id))
+    if New_SKU and New_order_price:
+        if Supplier_Goods_Info.objects.filter(SKU__contains=New_SKU):
+            Msg['msg_2'] = '该店内码已存在'
+            return render(request, 'Supplier_Goods_Info.html',
+                          {'goods_lists': goods_lists, 'Msg': Msg,
+                           'Edit_id': Edit_id, 'All_request': request.POST,
+                           'goods_len': goods_len, 'paginator': paginator})
+        else:
+            Supplier_Goods_Info.objects.create(SKU=New_SKU, order_price=New_order_price)
+    # 排序
+    SKU_Sort = request.POST.get("SKU_Sort")
+    order_price_Sort = request.POST.get("order_price_Sort")
+    Multistage_query = request.POST.get("Multistage_query")
+    Level_1 = request.POST.get("Level_1")
+    Level_2 = request.POST.get("Level_2")
+
+    sort_list = []
+    if Multistage_query:
+        if Level_1:
+            sort_list.append(Sort_condition(Level_1))
+            if Level_2:
+                sort_list.append(Sort_condition(Level_2))
+    else:
+        if Sort_condition(SKU_Sort):
+            sort_list.append(Sort_condition(SKU_Sort))
+        elif Sort_condition(order_price_Sort):
+            sort_list.append(Sort_condition(order_price_Sort))
+    goods_lists = goods_lists.order_by(*sort_list)
+    if Page_id != 1:
+        goods_lists = Record.Supplier_Goods_Info_Goods
+        request.POST = Record.Supplier_Goods_Info_Request
+    else:
+        Record.Supplier_Goods_Info_Request = request.POST
+        Record.Supplier_Goods_Info_Goods = goods_lists
+    goods_len = len(goods_lists)
+    # print('POST内容: ', request.POST)
+    paginator = Paginator(goods_lists, 10)
+    goods_lists = paginator.page(Page_id)
+    return render(request, 'Supplier_Goods_Info.html',
+                  {'goods_lists': goods_lists, 'Edit_id': Edit_id, 'All_request': request.POST,
+                   'goods_len': goods_len, 'Edit_object': Edit_object, 'paginator': paginator})
